DSEGOG-267 Checkpoint of validity checks refactor
