+++ operationsgateway_api/src/models.py
@@ -37,10 +37,10 @@ class ImageModel(BaseModel):
 
 
 class WaveformModel(BaseModel):
-    # Path is optional as we need it when ingesting waveforms (so we know where to
-    # Optional[Union[store it) but don't want to display it when a user, Any]] is
-    # retrieving a waveform. Setting `exclude=True` inside `Field()` ensure it's not
-    # displayed when returned as a response
+    # Path is optional as we need it when ingesting waveforms (so we know where to store
+    # it) but don't want to display it when a user is retrieving a waveform. Setting
+    # `exclude=True` inside `Field()` ensures it's not displayed when returned as a
+    # response
     path: Optional[str] = Field(None, exclude=True)
     x: Optional[Union[List[float], Any]]
     y: Optional[Union[List[float], Any]]
